+++ lib/common_ops/src/get_iou/get_iou.cpp
@@ -9,7 +9,7 @@ All Rights Reserved 2020.
 void get_iou(at::Tensor proposals_idx_tensor, at::Tensor proposals_offset_tensor, at::Tensor instance_labels_tensor, at::Tensor instance_pointnum_tensor, at::Tensor proposals_iou_tensor, int nInstance, int nProposal){
     int *proposals_idx = proposals_idx_tensor.data_ptr<int>();
     int *proposals_offset = proposals_offset_tensor.data_ptr<int>();
-    long *instance_labels = instance_labels_tensor.data_ptr<long>();
+    int *instance_labels = instance_labels_tensor.data_ptr<int>();
     int *instance_pointnum = instance_pointnum_tensor.data_ptr<int>();
 
     float *proposals_iou = proposals_iou_tensor.data_ptr<float>();
